HP & Jump power-ups
HP power-up works as intented, some pretty annoying bugs currently with the jump power-up: can't get any kinds of timers to work (to reset the jump force back to normal) and the item can sometimes be picked multiple times.

Next I'll try to create a completely new controller that is attached to the player which tracks all of the power-ups.